Move MyTrainable out.
